backend/pyproject.toml
[project]
name = "docpod-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "python-multipart>=0.0.9",
  "uvicorn[standard]>=0.30.0",
  "pydantic[email]>=2.7.0",
  "pydantic-settings>=2.3.0",
  "SQLAlchemy>=2.0.30",
  "asyncpg>=0.29.0",
  "alembic>=1.13.0",
  "python-jose[cryptography]>=3.3.0",
  "bcrypt>=4.0.1,<4.1.0",
  "passlib>=1.7.4",
  "celery>=5.4.0",
  "redis>=5.0.0",
  "structlog>=24.1.0",
  "greenlet>=3.3.1",
  "litellm>=1.61.0",
  "PyYAML>=6.0.2",
  "Jinja2>=3.1.4",
  "httpx>=0.27.0",
  "trafilatura>=1.12.2",
  "pypdf>=5.1.0",
  "python-docx>=1.1.2",
  "reportlab>=4.2.0",
]

[tool.dependencies]
dependency-groups = {dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "httpx>=0.27.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0",
]}
